refactor: reorganiza níveis de componentes
